+++ tests/concurrency/test_acquire_concurrency_slots.py
@@ -3,7 +3,7 @@ from unittest import mock
 import uuid
 
 from prefect.client.schemas.responses import MinimalConcurrencyLimitResponse
-from prefect.concurrency import acquire_concurrency_slots
+from prefect.concurrency.asyncio import acquire_concurrency_slots
 
 
 async def test_calls_increment_client_method():
